chore: extract coder bin handling to its own file
